+++ perception/object_merger/src/association/solver/successive_shortest_path.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/object_merger/association/solver/ssp.hpp"
+#include "autoware_object_merger/association/solver/ssp.hpp"
 
 #include <algorithm>
 #include <cassert>
